+++ CHANGELOG.md
@@ -9,6 +9,7 @@ I will do my best to guarantee that this project adheres to [Semantic Versioning
 
 ### Fixed
 - `NWTCPSocket` will not crash if `disconnect` is called before `connectTo`.
+- Hashing of `IPv4Address` might overflow on 32bit machines.
 
 ## 0.7.2
 ### Fixed